+++ docs/release_notes.rst
@@ -1,10 +1,25 @@
 Release Notes
 =============
 
-v4.5.0
+v6.0.0
 ------
 
 * Cache authentication credentials for performance.
+
+v5.0.0
+------
+
+* Default `SwiftPath.temp_url` to using `inline=True`, since that's the general
+  expectation for how we've used Apache filer / generally you expect URLs to
+  render in-browser rather than default to attachment.
+
+
+v4.5.0
+------
+
+* Added `SwiftPath.temp_url` for generating temporary object URLs
+* Added environment variable ``OS_TEMP_URL_KEY`` for pulling in default temp url key
+>>>>>>> origin/master
 
 v4.4.1
 ------